# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API with Stripe Payments",
        "description": "A comprehensive e-commerce API with Stripe payment integration, real-time WebSocket features, and comprehensive order management",
        "contact": {
            "email": "admin@ecommerce-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://e-commerce-api-new.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/stripe/webhook": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Stripe webhook endpoint",
                "description": "Handles Stripe webhook events to automatically update order payment statuses. This endpoint is called by Stripe when payment events occur and requires proper webhook signature verification.",
                "operationId": "stripeWebhook",
                "requestBody": {
                    "description": "Stripe webhook payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "description": "Event type",
                                        "type": "string",
                                        "example": "payment_intent.succeeded"
                                    },
                                    "data": {
                                        "properties": {
                                            "object": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Payment intent ID",
                                                        "type": "string",
                                                        "example": "pi_3SOMF4EFDrXcJGWZ18qFfsrp"
                                                    },
                                                    "amount": {
                                                        "description": "Amount in cents",
                                                        "type": "integer",
                                                        "example": 4999
                                                    },
                                                    "currency": {
                                                        "description": "Currency code",
                                                        "type": "string",
                                                        "example": "usd"
                                                    },
                                                    "status": {
                                                        "description": "Payment status",
                                                        "type": "string",
                                                        "example": "succeeded"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "payment_succeeded": {
                                    "summary": "Payment Intent Succeeded",
                                    "value": {
                                        "type": "payment_intent.succeeded",
                                        "data": {
                                            "object": {
                                                "id": "pi_3SOMF4EFDrXcJGWZ18qFfsrp",
                                                "amount": 4999,
                                                "currency": "usd",
                                                "status": "succeeded"
                                            }
                                        }
                                    }
                                },
                                "payment_failed": {
                                    "summary": "Payment Intent Failed",
                                    "value": {
                                        "type": "payment_intent.payment_failed",
                                        "data": {
                                            "object": {
                                                "id": "pi_3SOMF4EFDrXcJGWZ18qFfsrp",
                                                "amount": 4999,
                                                "currency": "usd",
                                                "status": "requires_payment_method"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook processed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid webhook signature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid webhook signature"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found for payment intent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "warning"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order not found for payment intent"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}